refactor(2): commands.js clean-up, create new separate files for commands in commands.js
