+++ src/Commands/jokes.js
@@ -1,0 +1,64 @@
+const { Commands, Command, RequiredArg } = require('../commands');
+
+
+//Joke commands
+
+Commands.hi = new Command("Says hi to you", (message, args) => {
+    message.reply("Hi.")
+}, "Joke")
+
+/*Commands.allutf8 = new Command("Get all existing characters in , (message, args) => {
+    if (message.author.id != "621307633718132746") {
+        throw("You are NOT using this.")
+    }
+    let msg = ""
+    for (let n = 0; n <= 0xFFFF; n++) {
+        msg = msg + n + ": " + String.fromCharCode(n) + ", "
+        if ((n % 50) == 0 && n > 0) {
+            message.channel.send(msg)
+            msg = ""
+        }
+    }
+}) removed due to it obviously spamming, will not be re-enabled unless necessary*/
+
+Commands.annoy = new Command("Annoys the person you want", (message, args) => {
+    message.channel.send(args[0].slice(0, 1900) + " you suck")
+}, "Joke", [new RequiredArg(0, "You gotta give me someone dumdum", "person")])
+
+Commands.comfort = new Command("Comforts the person you want", (message, args) => {
+    message.channel.send(args[0].slice(0, 1900) + " you don't suck")
+}, "Joke", [new RequiredArg(0, "You gotta give me someone dumdum", "person")])
+
+Commands.say = new Command("Says whatever you want", (message, args) => {
+    if (!args.join(" ")) { args[0] = "** **" }
+    message.channel.send(args.join(" ").slice(0, 1900))
+    message.delete()
+}, "Joke", [new RequiredArg(0, "I can't just say nothing, can I?", "...text")])
+
+Commands.hentai = new Command("Totally sends you hentai", (message, args) => {
+    message.channel.send("No..just no..")
+}, "Joke")
+
+Commands.secret = new Command("Sends you the secret of life", (message, args) => {
+    message.author.send("The sercret of life is...\n" +
+        "||⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣠⣤⣤⣤⣤⣤⣶⣦⣤⣄⡀⠀⠀⠀⠀⠀⠀⠀⠀||\n" +
+        "||⠀⠀⠀⠀⠀⠀⠀⠀⢀⣴⣿⡿⠛⠉⠙⠛⠛⠛⠛⠻⢿⣿⣷⣤⡀⠀⠀⠀⠀⠀||\n" +
+        "||⠀⠀⠀⠀⠀⠀⠀⠀⣼⣿⠋⠀⠀⠀⠀⠀⠀⠀⢀⣀⣀⠈⢻⣿⣿⡄⠀⠀⠀⠀||\n" +
+        "||⠀⠀⠀⠀⠀⠀⠀⣸⣿⡏⠀⠀⠀⣠⣶⣾⣿⣿⣿⠿⠿⠿⢿⣿⣿⣿⣄⠀⠀⠀||\n" +
+        "||⠀⠀⠀⠀⠀⠀⠀⣿⣿⠁⠀⠀⢰⣿⣿⣯⠁⠀⠀⠀⠀⠀⠀⠀⠈⠙⢿⣷⡄⠀||\n" +
+        "||⠀⠀⣀⣤⣴⣶⣶⣿⡟⠀⠀⠀⢸⣿⣿⣿⣆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣿⣷⠀||\n" +
+        "||⠀⢰⣿⡟⠋⠉⣹⣿⡇⠀⠀⠀⠘⣿⣿⣿⣿⣷⣦⣤⣤⣤⣶⣶⣶⣶⣿⣿⣿⠀||\n" +
+        "||⠀⢸⣿⡇⠀⠀⣿⣿⡇⠀⠀⠀⠀⠹⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠃⠀||\n" +
+        "||⠀⣸⣿⡇⠀⠀⣿⣿⡇⠀⠀⠀⠀⠀⠉⠻⠿⣿⣿⣿⣿⡿⠿⠿⠛⢻⣿⡇⠀⠀||\n" +
+        "||⠀⣿⣿⠁⠀⠀⣿⣿⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸⣿⣧⠀⠀||\n" +
+        "||⠀⣿⣿⠀⠀⠀⣿⣿⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸⣿⣿⠀⠀||\n" +
+        "||⠀⣿⣿⠀⠀⠀⣿⣿⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸⣿⣿⠀⠀||\n" +
+        "||⠀⢿⣿⡆⠀⠀⣿⣿⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸⣿⡇⠀⠀||\n" +
+        "||⠀⠸⣿⣧⡀⠀⣿⣿⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣿⣿⠃⠀⠀||\n" +
+        "||⠀⠀⠛⢿⣿⣿⣿⣿⣇⠀⠀⠀⠀⠀⣰⣿⣿⣷⣶⣶⣶⣶⠶⠀⢠⣿⣿⠀⠀⠀||\n" +
+        "||⠀⠀⠀⠀⠀⠀⠀⣿⣿⠀⠀⠀⠀⠀⣿⣿⡇⠀⣽⣿⡏⠁⠀⠀⢸⣿⡇⠀⠀⠀||\n" +
+        "||⠀⠀⠀⠀⠀⠀⠀⣿⣿⠀⠀⠀⠀⠀⣿⣿⡇⠀⢹⣿⡆⠀⠀⠀⣸⣿⠇⠀⠀⠀||\n" +
+        "||⠀⠀⠀⠀⠀⠀⠀⢿⣿⣦⣄⣀⣠⣴⣿⣿⠁⠀⠈⠻⣿⣿⣿⣿⡿⠏⠀⠀⠀⠀||\n" +
+        "||⠀⠀⠀⠀⠀⠀⠀⠈⠛⠻⠿⠿⠿⠿⠋⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀||\n" +
+        "|| amogus. ||")
+}, "Joke")
